move projectdetails to separate file
